leo-electrical-django/requirements.txt

Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
whitenoise==6.7.0
